+++ CHANGELOG.md
@@ -286,3 +286,4 @@
 * Define `yoklKL'yokl` operator
 * Define `yokl'yoklKL` operator
 * Define `constant` expression
+* Define `Fastening` type family